requirements.txt
aiohttp==3.7.4.post0
aiopg==1.3.1
async-timeout==3.0.1
attrs==21.2.0
chardet==4.0.0
flake8==3.9.2
idna==3.2
mccabe==0.6.1
multidict==5.1.0
mypy==0.910
mypy-extensions==0.4.3
peewee==3.14.4
peewee-async==0.7.1
psycopg2-binary==2.9.1
pycodestyle==2.7.0
pyflakes==2.3.1
python-dotenv==0.19.0
toml==0.10.2
typing-extensions==3.10.0.2
yarl==1.6.3
